chore(m0): move ESC/I-2 sources into src/esci2/ subfolder
