+++ tools/extract-test-jpeg.ts
@@ -9,8 +9,8 @@
 import fs from "node:fs";
 import path from "node:path";
 import os from "node:os";
-import { startScanSession } from "../src/scanner.js";
-import { FakeTlsSocket } from "../src/test-support/fake-tls-socket.js";
+import { startScanSession } from "../src/esci2/scanner.js";
+import { FakeTlsSocket } from "../src/esci2/test-support/fake-tls-socket.js";
 
 const CAPTURE = "tools/frida-capture/captures/2026-04-24T08-56-07-adf-1p-simplex.jsonl";
 const OUT = "test-fixtures/sample-page.jpg";
